💡 Configured idea service to use HTTP accessor and added new Forbidden exception

+++ src/backend/RateIdeas.Application/Users/Commands/DeleteUser/DeleteUser.cs
@@ -10,7 +10,10 @@ public class DeleteUserCommandHandler(IRepository<User> repository,
     public async Task<bool> Handle(DeleteUserCommand request, CancellationToken cancellationToken)
     {
         var entity = await repository.SelectAsync(entity => entity.Id.Equals(HttpContextHelper.GetUserId ?? 0))
-            ?? throw new UnAuthenticationException("Authentication has not been completed");
+            ?? throw new AuthenticationException("Authentication has not been completed");
+
+        if (entity.Role.Equals(Roles.SuperAdmin))
+            throw new ForbiddenExistException("Deleting a SuperAdmin is forbidden.");
 
         if (entity.Image is not null)
             await mediator.Send(new DeleteAssetCommand(entity.ImageId), cancellationToken);
