+++ src/backend/RateIdeas.Application/Users/Commands/DeleteUser/DeleteUserById.cs
@@ -14,6 +14,9 @@ public class DeleteUserByIdCommandHandler(IRepository<User> repository,
         var entity = await repository.SelectAsync(entity => entity.Id.Equals(request.Id))
             ?? throw new NotFoundException($"{nameof(User)} is not found by ID: {request.Id}");
 
+        if (entity.Role.Equals(Roles.SuperAdmin))
+            throw new ForbiddenExistException("Deleting a SuperAdmin is forbidden.");
+
         if (entity.Image is not null)
             await mediator.Send(new DeleteAssetCommand(entity.ImageId), cancellationToken);
